+++ frontend/src/App.js
@@ -10,6 +10,7 @@ import PressKit from "./pages/PressKit";
 import RarePreview from "./pages/RarePreview";
 import Rituals from "./pages/Rituals";
 import Journal from "./pages/Journal";
+import Secrets from "./pages/Secrets";
 import Conquest from "./pages/Conquest";
 import Duel from "./pages/Duel";
 import Bingo from "./pages/Bingo";
@@ -47,6 +48,7 @@ function App() {
           <Route path="/leaderboard" element={<Leaderboard />} />
           <Route path="/rituals" element={<Rituals />} />
           <Route path="/journal" element={<Journal />} />
+          <Route path="/secrets" element={<Secrets />} />
           <Route path="/conquest" element={<Conquest />} />
           <Route path="/d/:code" element={<Duel />} />
           <Route path="/bingo" element={<Bingo />} />